refactor: extract normalizeErrorCode and parseRetryCount to shared integrationsDiagnostics helper
Agent-Logs-Url: https://github.com/abell35213/adc-mvp/sessions/e5bc301d-2da7-4458-9b92-2f9303be84ba
